initrdscripts/rootfs: Use udev by-state partition links paths
When root is passed as a LABEL argument, we use the by-state udev link. For the
case of flasher images, we fallback to using the rootfs label if the by-state
symlink is unavailable.

The special case for flasher images is needed because while Grub based
flashers are able to pass the UUID of the root filesystem on the boot
media, U-Boot based flashers cannot and need to pass the root label.

We also move the /run mountpoint before switching to the new rootfs so
we are able to access the initramfs logs from a running balenaOS.

This commit is a re-work of previous work by Andrei Gherzan.

Signed-off-by: Alex Gonzalez <alexg@balena.io>