systemd: Add by-state udev links for all partitions
When a partition filesystem label is detected, udev checks whether the
device belongs to the same disk as the root partitition passed in the
kernel command line by the bootloader. Only if it does, it creates a by-state
link to it.

By using this by-state links we avoid filesystem label clashes as the
system will always use partitions in the same drive as root (as
instructed by the bootloader who is the source of truth).

Change-type: patch
Changelog-entry: Create udev state symlinks for all partitions
Signed-off-by: Alex Gonzalez <alexg@balena.io>